Updated data types

+++ src/schema/sensors.rs
@@ -22,8 +22,8 @@ pub struct Model {
     pub com_type: i32,
     #[sea_orm(column_name = "comLast")]
     pub com_last: Option<DateTime>,
-    #[sea_orm(column_name = "currentTemp", column_type = "Double", nullable)]
-    pub current_temp: Option<f64>,
+    #[sea_orm(column_name = "currentTemp", column_type = "Float", nullable)]
+    pub current_temp: Option<f32>,
     #[sea_orm(column_name = "currentHumid")]
     pub current_humid: Option<i32>,
     pub presence: Option<bool>,
